Move event-based solve from physics.py to module
- It can be a little hard to draw the line between physics and
  the equations that predict collisions, but I think this is the rights
  strategy given the limited utility of these functinos outside of the
  event_based solver.
